+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prairielearn/cache
 
+## 2.0.5
+
+### Patch Changes
+
+- a8438ff: Upgrade all JavaScript dependencies
+- Updated dependencies [a8438ff]
+- Updated dependencies [f33f309]
+  - @prairielearn/logger@2.0.4
+  - @prairielearn/sentry@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes